Refactor ephys features into 3 modules: subthreshold, spike and stim _features. Corrspondingly split tests.
