+++ ipfx/bin/pipeline_from_nwb_file.sh
