+++ ipfx/ephys_data_set.py
@@ -1,4 +1,5 @@
 import logging
+import numpy as np
 from ipfx.stimulus import StimulusOntology
