+++ ipfx/error.py
@@ -1,0 +1,4 @@
+
+class FeatureError(Exception):
+    """Generic Python-exception-derived object raised by feature detection functions."""
+    pass
