+++ run_tests_with_docker.sh
